requirements.txt
click==8.1.3
colorama==0.4.4
Flask==2.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.3
PyJWT==2.4.0
typing-extensions==4.2.0
Werkzeug==2.1.2
zipp==3.8.0
